+++ app/components/LoginPanel/LoginPanel.js
@@ -36,9 +36,25 @@ export default withRouter(class LoginPanel extends Component {
           store.dispatch(LOGIN_CREDENTIALS_SUCCESS(val));
           this.props.router.replace('authenticated');
         }).catch((err) => {
+        console.log(err);
           store.dispatch(LOGIN_CREDENTIALS_FAILURE(err));
         });
     });
+  }
+
+  parseFormData() {
+    const args = arguments;
+    let body;
+
+    for (let i = 0, length = args.length; i < length; i++) {
+      const argExists = args[i].split('=')[1].length >= 1;
+
+      if (argExists) {
+        body = `${body ? `${body}&` : ''}${args[i]}`;
+      }
+    }
+
+    return body;
   }
 
   /**
@@ -46,36 +62,47 @@ export default withRouter(class LoginPanel extends Component {
    * @returns {Promise}
    */
   submitLogin() {
-    const username = document.getElementById('usernameInput').value;
-    const password = document.getElementById('passwordInput').value;
-    
+    const email = `email=${document.getElementById('usernameInput').value}`;
+    const password = `password=${document.getElementById('passwordInput').value}`;
+
     // do some client-side validation
-    
+
+    const parsed = this.parseFormData(email, password);
+
     return new Promise((resolve, reject) => {
       const xhr = new XMLHttpRequest();
       
       xhr.open('POST', encodeURI('http://localhost:3030/login'));
-      
+      xhr.setRequestHeader("Content-type", "application/x-www-form-urlencoded");
+
       xhr.onload = () => {
-        if (xhr.status !== 200) {
+        const res = xhr.response;
+
+        console.log(res);
+        
+        if (!res || xhr.status !== 200) {
           reject(`${xhr.status}: ${xhr.statusText}`);
         }
         
-        resolve(xhr.response);
+        resolve(res);
       };
       
       store.dispatch(LOGIN_CREDENTIALS_REQUEST());
-      
-      xhr.send();
+      xhr.send(parsed);
     });
   }
 
   render() {
     let loader;
     let containerClasses;
+    let errorMessage;
 
     if (this.props.login.authInProgress) {
-      loader = <div className={d.loader}></div>;
+      loader = <div className={d.loader}>Woooooooooooo loading</div>;
+    }
+    
+    if (this.props.login.failedLogin) {
+      errorMessage = <h4>Invalid credentials</h4>;
     }
 
     if (this.props.customClass) {
@@ -110,7 +137,7 @@ export default withRouter(class LoginPanel extends Component {
           </form>
 
           {
-              this.props.login.authInProgress
+              errorMessage
           }
 
           {