+++ server/auth/localStrategy.js
@@ -1,9 +1,0 @@
-/**
- * Created by jahansj on 05/11/2016.
- */
-const passport = require('passport');
-const localStrategy = require('passport-local');
-
-passport.use(new localStrategy((username, password, done) => {
-  
-}));